% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bays_2009_error.R
\name{bays_2009_error}
\alias{bays_2009_error}
\title{bays_2009_error}
\usage{
bays_2009_error(X, Tg = 0)
}
\arguments{
\item{X}{response vector radians}

\item{Tg}{target orientations in radians}
}
\description{
\code{bays_2009_error}
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
Bays PM, Catalao RFG & Husain M. The precision of visual working
}
