% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bays_2009_fit_df.R
\name{bays_2009_fit_df}
\alias{bays_2009_fit_df}
\title{bays_2009_fit_df}
\usage{
bays_2009_fit_df(d, id.var = "id", tar.var = "target",
  res.var = "response", nt.vars = NULL)
}
\arguments{
\item{d}{dataframe}

\item{id.var}{id variable string}

\item{tar.var}{target variable}

\item{res.var}{response variable}

\item{nt.vars}{optional non target vars}
}
\description{
\code{bays_2009_fit_df}
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
Bays PM, Catalao RFG & Husain M. The precision of visual working
}
