% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bays_2009_function.R
\name{bays_2009_function}
\alias{bays_2009_function}
\title{bays_2009_function}
\usage{
bays_2009_function(X, Tg, NT = replicate(NROW(X), 0), B_start = NULL)
}
\arguments{
\item{X}{response vector radians}

\item{Tg}{target orientations in radians}

\item{NT}{non target matrix}

\item{B_start}{starting params}
}
\description{
\code{bays_2009_function}
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
Bays PM, Catalao RFG & Husain M. The precision of visual working
}
