% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bays_2009_likelihood.R
\name{bays_2009_likelihood}
\alias{bays_2009_likelihood}
\title{bays_2009_likelihood}
\usage{
bays_2009_likelihood(B, X, Tg, NT = replicate(NROW(X), 0))
}
\arguments{
\item{B}{params}

\item{X}{response vector radians}

\item{Tg}{target orientations in radians}

\item{NT}{non target matrix}
}
\description{
\code{bays_2009_likelihood}
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
Bays PM, Catalao RFG & Husain M. The precision of visual working
}
