% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bays_2009_error_df.R
\name{bays_2009_df_error}
\alias{bays_2009_df_error}
\title{bays_2009_df}
\usage{
bays_2009_df_error(d, id.var = "id", tar.var = "target",
  res.var = "response")
}
\description{
\code{bays_2009_df}
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
Bays PM, Catalao RFG & Husain M. The precision of visual working
}
