% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bays_2009_fit.R
\name{bays_2009_fit}
\alias{bays_2009_fit}
\title{bays_2009_df}
\usage{
bays_2009_fit(X, Tg, NT = replicate(NROW(X), 0), return.ll = TRUE)
}
\arguments{
\item{X}{response vector radians}

\item{Tg}{target orientations in radians}

\item{NT}{non target matrix}

\item{return.ll}{logic return log liklihood}
}
\description{
\code{bays_2009_df}
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
Bays PM, Catalao RFG & Husain M. The precision of visual working
}
