% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_bias.R
\name{cir_bias}
\alias{cir_bias}
\title{Circular bias}
\usage{
cir_bias(x)
}
\arguments{
\item{x}{A vector of radians in the range -pi to pi}
}
\description{
#' @aliases cir_mean
}
\details{
\code{cir_sd} takes of vector of radians and returns the circular equivalent of the standard deviation
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
}
