% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_mean.R
\name{cir_mean}
\alias{cir_mean}
\title{Circular mean}
\usage{
cir_mean(x)
}
\arguments{
\item{x}{A vector of radians in the range -pi to pi}
}
\description{
#' @aliases cir_bias
}
\details{
\code{cir_sd} takes of vector of radians and returns the circular equivalent of the standard deviation
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
}
