% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_precision.R
\name{cir_precision}
\alias{cir_precision}
\title{Circular Precision}
\usage{
cir_precision(X, Tg = 0)
}
\arguments{
\item{X}{response vector radians}

\item{Tg}{target orientations in radians}
}
\description{
\code{cir_sd} takes of vector of radians and returns the circular equivalent of the standard deviation
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
}
