% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_sd.R
\name{cir_sd}
\alias{cir_sd}
\title{Circular standard deviation}
\usage{
cir_sd(x)
}
\arguments{
\item{x}{A vector of radians in the range -pi to pi}
}
\description{
\code{cir_sd} takes of vector of radians and returns the circular equivalent of the standard deviation
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
}
