% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k2sd.R
\name{k2sd}
\alias{k2sd}
\title{k2sd}
\usage{
k2sd(K)
}
\arguments{
\item{K}{concentration}
}
\description{
\code{k2sd}
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
Ref: Topics in Circular Statistics, S. R. Jammalamadaka & A. Sengupta
}
