% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logspace.R
\name{logspace}
\alias{logspace}
\title{k2sd}
\usage{
logspace(a, b, n)
}
\arguments{
\item{a}{a}

\item{b}{b}

\item{n}{n}
}
\description{
\code{k2sd}
}
\references{
http://r.789695.n4.nabble.com/logarithmic-seq-tp900431p900433.html
}
