% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomvonmises.R
\name{randomvonmises}
\alias{randomvonmises}
\title{Random vonmises}
\usage{
randomvonmises(N, MU, K)
}
\arguments{
\item{N}{number of random samples to generate}

\item{MU}{mean of the distribution}

\item{K}{concentration}
}
\description{
\code{randomvonmises}
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
Best, D. and Fisher, N. (1979). Applied Statistics, 24, 152-157.
}
