% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd2k.R
\name{sd2k}
\alias{sd2k}
\title{sd2k}
\usage{
sd2k(S)
}
\arguments{
\item{S}{std}
}
\description{
\code{sd2k}
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
Ref: Topics in Circular Statistics, S. R. Jammalamadaka & A. Sengupta
}
