% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonmisespdf.R
\name{vonmisespdf}
\alias{vonmisespdf}
\title{Random vonmises}
\usage{
vonmisespdf(X, MU, K)
}
\arguments{
\item{MU}{mean of the distribution}

\item{K}{concentration}
}
\description{
\code{randomvonmises}
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
Best, D. and Fisher, N. (1979). Applied Statistics, 24, 152-157.
}
