% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{Random vonmises}
\usage{
wrap(Y, bound = pi)
}
\arguments{
\item{Y}{a vec}

\item{bound}{pi}
}
\description{
\code{randomvonmises}
}
\references{
Adapted from Matlab code by Paul Bays (https://www.paulbays.com/code.php)
Best, D. and Fisher, N. (1979). Applied Statistics, 24, 152-157.
}
